which variant and which batch of a vars file are counted with one type
`VarsNullValue` held the variant it is about as a `u64` and the two cases
about a batch held the batch as a `usize`, and the three say the same kind of
thing, which one of the file counted from 1. They are a `u64` now, as the
`coding` skill asks of a count that runs over the whole dataset: what a
`usize` holds depends on the target, 4295 million in wasm, and a file with
more variants than that is read there too.

The change is of a type and has no test of its own: nothing of what the three
messages say changes, and the compiler is what checks the callers, which
today are the tests of the two modules.

+++ crates/popnei/src/error.rs
@@ -322,7 +322,8 @@ pub enum Error {
     VarsNullValue {
         /// The name of the column.
         column: &'static str,
-        /// Which variant of the file it is, counted from 1.
+        /// Which variant of the file it is, counted from 1, over the whole
+        /// file and not inside its batch.
         var: u64,
     },
 
@@ -348,8 +349,11 @@ pub enum Error {
         "the batch {batch} of the vars file holds {found} variants and its entry of the `popnei_batches` key of the footer says {expected}"
     )]
     VarsBatchNumVars {
-        /// Which batch of the file it is, counted from 1.
-        batch: usize,
+        /// Which batch of the file it is, counted from 1. It is a `u64` as
+        /// the variant of [`Error::VarsNullValue`] is: both count over the
+        /// whole file, which a machine that counts to 4295 million reads
+        /// too.
+        batch: u64,
         /// How many variants it holds.
         found: usize,
         /// How many its entry says.
@@ -502,8 +506,11 @@ pub enum Error {
         "the batch {batch} of the vars file could not be read, so the file is damaged and has to be fetched or copied again: {problem}"
     )]
     VarsBatchNotRead {
-        /// Which batch of the file it is, counted from 1.
-        batch: usize,
+        /// Which batch of the file it is, counted from 1. It is a `u64` as
+        /// the variant of [`Error::VarsNullValue`] is: both count over the
+        /// whole file, which a machine that counts to 4295 million reads
+        /// too.
+        batch: u64,
         /// What arrow-rs said about it.
         problem: String,
     },
